==> 3_calc_fis_scores.Rmd <==
---
title: 'OHIBC: Calculate FIS scores'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/ohibc/src/templates/ohibc_hdr1.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 8, fig.height = 6, echo = TRUE, message = FALSE, warning = FALSE, fig.path = 'figs/')

dir_git <- '~/github/ohibc'
source(file.path(dir_git, 'src/R/common.R'))  ### an OHIBC specific version of common.R
dir_rgn <- file.path(dir_git, 'prep/regions')  ### github: general buffer region shapefiles
dir_anx <- file.path(dir_M, 'git-annex/bcprep')

### support scripts
source("~/github/ohibc/calc_ohibc/calc_scores_fxns.R")
```

# Summary

This script sources `functions.R` (the FIS model). 

This model has the following characteristics:

- Stock scores are weighted by proportional catch  
- A penalty for unassessed catch is applied to the status, equal to 1 minues the % of unassessed catch  
- Underfishing penalty is now at 0.66 instead of 0.8 (meaning F/Fmsy >= 0.66 & <= 1.2 is assigned a score of 1)

------

```{r source fisheries function}
#this runs the fisheries goal model and produces stock_scores.csv and fis_status.csv
source("fis_function.R")
```


## Stock scores and overall scores by region

I want to look at the stock scores for each region's stocks, and the average score. 

``` {r assemble_dataframe}
## plotting fishery catch weighting by region
stock_plot_df <- read_csv('output/stock_scores.csv') %>%
  mutate(weighted_score = score * rgn_ass_catch_prop) %>%
  group_by(rgn_name, year) %>%
  mutate(rgn_score = sum(weighted_score, na.rm = T)) 
```

We currently have an NA region due to missing Albacore Tuna data. The Albacore Tuna stock data does not have any region assigned to it. I could manuallyl assign this stock to the pacific offshore region but I want to wait until we have it from Karen. For now, you can imagine that data would be put into the offshore region.

The line thickness shows the proportional catch of each stock within a region. The thicker the line, the more that stock contributes to the final score.

``` {r plot_fis_stock_scores}

rgn_avg <- stock_plot_df %>%
  select(year, rgn_score, rgn_name) %>%
  distinct()

ggplot(stock_plot_df, aes(x = year, y = score)) +
    ggtheme_plot() +
    geom_line(aes(group = stock_id, color = stock_id, size=rgn_ass_catch_prop),
              lineend = 'round', alpha = 0.8) +
    labs(color = 'Stock ID',
         y     = 'Stock Score') +
    guides(colour = guide_legend(override.aes = list(size = 3)),
           size = 'none') +
    facet_wrap(~rgn_name) +
    geom_line(data = rgn_avg, aes(x = year, y = rgn_score, group = rgn_name), color = "black", lwd = 1)

```


Looking at catch over time by region. Need to plot each stock as a line but differentiate between assessed and unassessed.

```{r, fig.width = 10, fig.height = 8}

dfo_catch <- read_csv("output/dfo_catch.csv") %>%
  mutate(assessed = ifelse(is.na(stockid), "unassessed", "assessed")) %>%
  select(year, rgn_name, species, catch, assessed) %>%
  distinct()

ggplot(dfo_catch, aes(x = year, y = catch, color = species)) +
  geom_line(aes(linetype = assessed)) +
  facet_wrap(~rgn_name, scales = "free") +
  theme_bw()

```

#Fisheries status scores

Here are the status scores for each region over time.

``` {r plot_fis_status_scores}

status <- read_csv("output/fis_status.csv")

ggplot(status, aes(x = year, y = score, group = rgn_name)) +
    ggtheme_plot() +
    geom_line(aes(color = rgn_name), alpha = 0.8) +
    labs(color = 'Region',
         y     = 'Fisheries Status') +
    ggtitle("Fisheries status in British Columbia")

```

Let's look at proportion of unassessed catch over time for each region.

```{r}
unass <- read_csv("output/rgn_catch_summary.csv")
ggplot(unass, aes(x = year, y = (unass_catch_prop * 100), color = rgn_name)) +
  ggtheme_plot() +
  geom_line() +
  labs(title = "Percent of unassessed catch",
       y = "Percent",
       x = "Year",
       color = "Region")

```
